Increased floating point precision

+++ src/math/toughness.rs
@@ -1,6 +1,6 @@
 // FIXME: This also shouldnt be in 1.0.0
 #[allow(dead_code)]
-fn calculate_level_multiplier(level: u8) -> f32 {
+fn calculate_level_multiplier(level: u8) -> f64 {
     match level {
         1 => 54.0000,
         2 => 58.0000,
